// components/ProjectStatusManager/0.1.0.0/src/ts/layer3/ProjectStatusManagerModel.interface.ts
/**
 * ProjectStatusManagerModel - ProjectStatusManager Component Model Interface
 * Web4 pattern: Component model following auto-discovery patterns
 */

import { Model } from './Model.interface.js';

export interface ProjectStatusManagerModel extends Model {
  uuid: string;
  name: string;
  origin: string;
  definition: string;
  createdAt: string;
  updatedAt: string;
}